requirements.txt
discord.py
ephem
python-dotenv==1.0.1